skips N pixels, everything works, but it could look a bit better

+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: awoimbee <awoimbee@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/10/23 20:34:49 by awoimbee          #+#    #+#             */
-/*   Updated: 2019/02/27 19:25:43 by awoimbee         ###   ########.fr       */
+/*   Updated: 2019/02/27 21:21:39 by awoimbee         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,7 +39,7 @@ void			ft_memdel(void **ap);
 void			*ft_mempcpy(void *dst, const void *src, size_t n);
 void			ft_mem64set(uint64_t *mem, uint64_t data, size_t memlen);
 void			ft_mem32set(uint32_t *mem, uint32_t data, size_t memlen);
-void			ft_mem16set(uint64_t *mem, uint16_t data, size_t memlen);
+void			ft_mem16set(uint16_t *mem, uint16_t data, size_t memlen);
 
 /*
 **	LIBSTR
